Removed unneeded collider from elevator, improved finalLevel colliders
Removed colliders from most of the rubble, replacing the bigger piles' colliders with cube colliders, in order to prevent the player from getting stuck in it, and make it easier to navigate. Also changes the stairs from regular cube meshes to a plane mesh, allowing the player to walk up it rather than forcing them to constantly jump up.